requirements.txt
fastapi
uvicorn

SQLAlchemy
pymysql

python-dotenv
pydantic-settings

requests

python-jose
passlib[bcrypt]

watchfiles
